Put kernels in separate file, improve MMD tests
