+++ test/renate/shift/test_kernels.py
@@ -1,0 +1,39 @@
+# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# SPDX-License-Identifier: Apache-2.0
+import pytest
+import torch
+
+from renate.shift.kernels import RBFKernel
+
+
+@pytest.mark.parametrize("kernel", [RBFKernel()])
+@pytest.mark.parametrize(
+    "X1, X2",
+    [
+        (torch.randn(20, 2), torch.randn(20, 2)),
+        (torch.randn(10, 2), torch.randn(20, 2)),
+        (torch.randn(20, 2), torch.randn(10, 2)),
+    ],
+)
+def test_kernel_shapes(kernel, X1, X2):
+    K = kernel(X1, X2)
+    assert K.size() == (X1.size(0), X2.size(0))
+
+
+@pytest.mark.parametrize("kernel", [RBFKernel()])
+def test_kernel_shape_mismatch(kernel):
+    with pytest.raises(Exception):
+        kernel(torch.randn(10, 2), torch.randn(20, 3))
+
+
+def test_rbf_kernel_limits():
+    """Tests limit behavior of RBF kernel"""
+    X = torch.randn(10, 2)
+    # Small lengthscales should result in vanishing off-diagonal terms.
+    kernel = RBFKernel(lengthscale=1e-8)
+    K = kernel(X, X)
+    assert torch.allclose(K, torch.eye(10))
+    # High lengthscales should result in
+    kernel = RBFKernel(lengthscale=1e8)
+    K = kernel(X, X)
+    assert torch.allclose(K, torch.ones(10))
